+++ src/app/services/dimPlatformService.factory.js
@@ -1,56 +1,94 @@
 import angular from 'angular';
 import _ from 'underscore';
+import Rx from 'rxjs';
 
 angular.module('dimApp').factory('dimPlatformService', PlatformService);
 
 
-function PlatformService($rootScope, BungieAccountService, DestinyAccountService, SyncService, $state, $q, dimState) {
+function PlatformService($rootScope, $q, BungieUserApi, SyncService, OAuthTokenService, $state, toaster, loadingTracker) {
   let _platforms = [];
   let _active = null;
 
+  /**
+   * The platforms observable will update with the list of platforms. It won't ever update after that.
+   */
+  // TODO: this would probably be best as just a promise...
+  const platforms = Rx.Observable
+    .defer(() => {
+      const platformsPromise = getPlatforms();
+      loadingTracker.addPromise(platformsPromise);
+      return Rx.Observable.fromPromise(platformsPromise);
+    })
+    .publishBehavior() // Keep the value around after the promise has resolved
+    .refCount(); // Cache the response forever? Might not be what you want.
+
+  // This should probably just be a regular subject?
+  const activePlatform = new Rx.BehaviorSubject(null);
+
   const service = {
-    getPlatforms,
-    getActive,
-    setActive,
-    reportBadPlatform,
-    getPlatformMatching
+    getPlatforms: getPlatforms,
+    getActive: getActive,
+    setActive: setActive,
+    reportBadPlatform: reportBadPlatform,
+    platforms, // an observable of platform updates
+    activePlatform // an observable of the currently active platform
   };
 
   return service;
 
-  function getPlatformMatching(params) {
-    return _.find(_platforms, params);
-  }
-
-  /**
-   * @return {DestinyAccount[]}
-   */
-  // TODO: return a list of bungie accounts and associated destiny accounts?
   function getPlatforms() {
-    if (_platforms.length) {
-      return $q.resolve(_platforms);
+    const token = OAuthTokenService.getToken();
+    if (!token) {
+      // We're not logged in, don't bother
+      $state.go('login');
+      return $q.when();
     }
 
-    // TODO: wire this up with observables?
-    return BungieAccountService.getBungieAccounts()
-      .then((bungieAccounts) => {
-        if (!bungieAccounts.length) {
-          // We're not logged in, don't bother
-          $rootScope.$broadcast('dim-no-token-found');
-          return [];
-        }
+    if (token.bungieMembershipId) {
+      return BungieUserApi.getAccounts(token.bungieMembershipId)
+        .then(generatePlatforms)
+        .catch((e) => {
+          toaster.pop('error', 'Unexpected error getting accounts', e.message);
+          throw e;
+        });
+    } else {
+      // they're logged in, just need to fill in membership
+      // TODO: this can be removed after everyone has had a chance to upgrade
+      return BungieUserApi.getAccountsForCurrentUser()
+        .then((accounts) => {
+          const token = OAuthTokenService.getToken();
+          token.bungieMembershipId = accounts.bungieNetUser.membershipId;
+          OAuthTokenService.setToken(token);
 
-        // We only support one account now
-        const membershipId = bungieAccounts[0].membershipId;
-        return DestinyAccountService.getDestinyAccountsForBungieAccount(membershipId);
-      })
-      .then((destinyAccounts) => {
-        _platforms = destinyAccounts;
-        $rootScope.$broadcast('dim-platforms-updated', { platforms: _platforms });
-        return getActivePlatform();
-      })
-      .then(setActive)
-      .then(() => _platforms);
+          return accounts;
+        })
+        .then(generatePlatforms)
+        .catch((e) => {
+          toaster.pop('error', 'Unexpected error getting accounts', e.message);
+          throw e;
+        });
+    }
+  }
+
+  function generatePlatforms(accounts) {
+    _platforms = accounts.destinyMemberships.map((destinyAccount) => {
+      const account = {
+        id: destinyAccount.displayName,
+        type: destinyAccount.membershipType,
+        membershipId: destinyAccount.membershipId
+      };
+      account.label = account.type === 1 ? 'Xbox' : 'PlayStation';
+      return account;
+    });
+
+    // platforms.next(_platforms);
+
+    getActivePlatform()
+      .then((activePlatform) => {
+        setActive(activePlatform);
+      });
+
+    return _platforms;
   }
 
   function getActivePlatform() {
@@ -63,7 +101,7 @@ function PlatformService($rootScope, BungieAccountService, DestinyAccountService
         return _active;
       } else if (data && data.platformType) {
         const active = _.find(_platforms, (platform) => {
-          return platform.platformType === data.platformType;
+          return platform.type === data.platformType;
         });
         if (active) {
           return active;
@@ -84,23 +122,20 @@ function PlatformService($rootScope, BungieAccountService, DestinyAccountService
     if (platform === null) {
       promise = SyncService.remove('platformType');
     } else {
-      promise = SyncService.set({ platformType: platform.platformType });
+      promise = SyncService.set({ platformType: platform.type });
     }
 
-    dimState.active = platform;
+    activePlatform.next(_active);
     $rootScope.$broadcast('dim-active-platform-updated', { platform: _active });
     return promise;
   }
 
   // When we find a platform with no characters, remove it from the list and try something else.
   function reportBadPlatform(platform, e) {
-    // TODO: push this up to DestinyAccountService
-
     if (_platforms.length > 1) {
       _platforms = _platforms.filter((p) => p !== platform);
-      $rootScope.$broadcast('dim-platforms-updated', { platforms: _platforms });
+      platforms.next(_platforms);
       setActive(_platforms[0]);
-      $state.go('destiny1'); // try for another platform
     } else {
       // Nothing we can do
       throw e;
